Split location and hebrew date, remove tests_common
