+++ hdate/__init__.py
@@ -5,9 +5,10 @@ HDate calculates and generates a representation either in English or Hebrew
 of the Jewish calendrical date and times for a given location
 """
 
-from hdate.common import HebrewDate, Location
 from hdate.date import HDate
+from hdate.hebrew_date import HebrewDate
 from hdate.htables import HolidayTypes
+from hdate.location import Location
 from hdate.zmanim import Zmanim
 
 __all__ = ["HDate", "Zmanim", "HebrewDate", "Location", "HolidayTypes"]
