+++ hdate/date.py
@@ -11,31 +11,18 @@ from itertools import chain, product
 
 from hdate import converters as conv
 from hdate import htables
+from hdate.hebrew_date import HebrewDate
 from hdate.htables import HolidayTypes, Months
 
 _LOGGER = logging.getLogger(__name__)
 # pylint: disable=too-many-public-methods
-
-@dataclass
-class HebrewDate:
-    """Define a Hebrew date object."""
-
-    year: int
-    month: Months
-    day: int
-
-    def __post_init__(self):
-        self.month = (
-            self.month if isinstance(self.month, Months) else Months(self.month)
-        )
-
 
 
 class HDate:
     """
     Hebrew date class.
 
-    Represents a hebrew date in the form of da
+    Supports converting from Gregorian and Julian to Hebrew date.
     """
 
     def __init__(
